refactor: fix Remove not by creator

+++ app/Http/Controllers/TaskController.php
@@ -96,6 +96,8 @@ class TaskController extends Controller
 
     public function destroy(Task $task)
     {
+        $this->authorize('delete', [self::class]);
+
         $task->labels()->detach();
         $task->delete();
